Add scenarios for formatting cleanup and scenario events in JSON
